requirement.txt
Pillow
Azure-cognitiveservices-vision-customvision
python-dotenv